+++ audio-ui/src/debug/java/com/google/android/horologist/audio/ui/VolumeScreenUxGuide.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.horologist.audioui
+package com.google.android.horologist.audio.ui
 
 import android.graphics.Color.WHITE
 import android.graphics.Paint
