+++ sample/src/main/java/com/google/android/horologist/sample/media/MediaPlayerScreen.kt
@@ -24,8 +24,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusRequester
 import androidx.wear.compose.material.Scaffold
 import androidx.wear.compose.material.TimeText
-import com.google.android.horologist.audioui.VolumePositionIndicator
-import com.google.android.horologist.audioui.VolumeViewModel
+import com.google.android.horologist.audio.ui.VolumePositionIndicator
+import com.google.android.horologist.audio.ui.VolumeViewModel
 import com.google.android.horologist.compose.navscaffold.scrollableColumn
 import com.google.android.horologist.compose.pager.FocusOnResume
 import com.google.android.horologist.media.ui.screens.PlayerScreen
